Add pre-commit for linting
This pre-commit configuration primarily ensures that the YAML is
well-configured. The check-jsonschema is particularly helpful because it
validates both workflows and the composite action source against a
schema.

NOTE: turns out the github-actions schema doesn't know how to deal with
an action.yaml that has a runs.using="composite" field, instead of one
that points to a JavaScript file. Commenting out for now.