Use uv for uploads
The GitHub Action to which we previously delegated the PyPI uploads
cannot be called from a composite action when using trusted
environments. Since uv now has built-in support for building and
uploading packages, switch to uv for both steps, dropping use of both
the PyPA build tool and twine. This significantly simplifies the
action and removes the need for the `python-version` parameter.

In the interests of removing twine from the workflow entirely, this
drops the twine check step, which was previously called when uploading
was disabled. Per the twine documentation, that step only checked
whether the package long description can be rendered.